Introduce shared pkg for test result methods
