+++ test/e2e/tools/eks-anywhere-test-tool/pkg/constants/constants.go
@@ -7,7 +7,7 @@ const (
 	E2eIndividualTestLogGroup = "/eks-anywhere/test/e2e"
 	E2eArtifactsBucketEnvVar  = "E2E_ARTIFACTS_BUCKET"
 	FailedMessage             = "An e2e instance run has failed"
-	SuccessMEssage            = "Ec2 instance tests completed successfully"
+	SuccessMessage            = "Ec2 instance tests completed successfully"
 	FailedTestsFile           = "failed-tests.txt"
 	LogOutputFile             = "codebuild-log.txt"
 	CiProxyLogGroup           = "nginx-vcenter-proxy.log"
